Improve Flycheck's modeline
We now show error, warning, and info counts separated by spaces, with
highlighting and tooltips (though neither may be visible in all modeline
packages, e.g. SML doesn't show them).  We have to use a hack to get mouse
clicks to work because mouse-minor-mode-menu captures just "FlyC" instead of the
whole string, and then fails to match that against the mode line strings it
knows about.

Instead of showing FlyC for both "not checked" and "no errors", we now show
FlyC~ for "not checked" and FlyC for "checked, no errors".